refactor: PrismaProfileFollowRepository
